+++ packages/studio-graph/src/graph/utils/handleStatus.tsx
@@ -1,4 +1,6 @@
-export const handleStatus = (item, runtimeStatus) => {
+import { EdgeStatus, NodeStatus } from '../types';
+
+export const handleStatus = (item, runtimeStatus): NodeStatus | EdgeStatus => {
   const { id, label, __status } = item;
   /** 用户手动指定的样式，优先级第一 */
   /** 按照label 划分群组的样式，优先级第二 */